+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 ### Removed
 ### Fixed
 - Fixed a bug where the DefaultAccuracy metric would not put the inner metric in eval mode if the first call to reset was after the call to eval
+- Fixed a bug where trying to load a state dict in a different session to where it was saved didn't work properly
 
 ## [0.2.1] - 2018-09-11
 ### Added